+++ attribution/src/main/java/com/affise/attribution/referrer/ReferrerCallback.kt
@@ -1,0 +1,5 @@
+package com.affise.attribution.referrer
+
+fun interface ReferrerCallback {
+    fun getReferrer(callback: OnReferrerCallback)
+}