+++ internal/src/main/java/com/affise/attribution/internal/AffiseApiWrapper.kt
@@ -143,6 +143,12 @@ class AffiseApiWrapper(
             AffiseApiMethod.GET_REFERRER_URL_VALUE_CALLBACK ->
                 callGetReferrerUrlValue(api, map, result)
 
+            AffiseApiMethod.GET_REFERRER_ON_SERVER_CALLBACK ->
+                callGetReferrerOnServer(api, map, result)
+
+            AffiseApiMethod.GET_REFERRER_ON_SERVER_VALUE_CALLBACK ->
+                callGetReferrerOnServerValue(api, map, result)
+
             AffiseApiMethod.REGISTER_DEEPLINK_CALLBACK ->
                 callRegisterDeeplinkCallback(api, map, result)
 
@@ -518,6 +524,44 @@ class AffiseApiWrapper(
             result.error("api [${api.method}]: value not set")
         } else {
             Affise.getReferrerUrlValue(key) { referrer ->
+                val data = mapOf<String, Any?>(
+                    UUID to uuid,
+                    api.method to referrer,
+                )
+                callback?.invoke(api, data)
+            }
+            result.success(null)
+        }
+    }
+
+    private fun callGetReferrerOnServer(
+        api: AffiseApiMethod,
+        map: Map<String, *>,
+        result: InternalResult
+    ) {
+        val uuid = map.opt<String>(UUID)
+        Affise.getReferrerOnServer { referrer ->
+            val data = mapOf<String, Any?>(
+                UUID to uuid,
+                api.method to referrer,
+            )
+            callback?.invoke(api, data)
+        }
+        result.success(null)
+    }
+
+    private fun callGetReferrerOnServerValue(
+        api: AffiseApiMethod,
+        map: Map<String, *>,
+        result: InternalResult
+    ) {
+        val uuid = map.opt<String>(UUID)
+        val key = map.opt<String>(api)?.toReferrerKey()
+
+        if (key == null) {
+            result.error("api [${api.method}]: value not set")
+        } else {
+            Affise.getReferrerOnServerValue(key) { referrer ->
                 val data = mapOf<String, Any?>(
                     UUID to uuid,
                     api.method to referrer,