+++ module-status/src/main/java/com/affise/attribution/module/status/StatusModule.kt
@@ -6,18 +6,24 @@ import com.affise.attribution.converter.StringToKeyValueConverter
 import com.affise.attribution.executors.ExecutorServiceProviderImpl
 import com.affise.attribution.module.status.usecase.CheckStatusUseCase
 import com.affise.attribution.module.status.usecase.CheckStatusUseCaseImpl
+import com.affise.attribution.module.status.usecase.ReferrerUseCase
+import com.affise.attribution.module.status.usecase.ReferrerUseCaseImpl
 import com.affise.attribution.modules.AffiseModule
 import com.affise.attribution.modules.OnKeyValueCallback
 import com.affise.attribution.modules.exceptions.AffiseModuleError
 import com.affise.attribution.network.HttpClient
 import com.affise.attribution.parameters.factory.PostBackModelFactory
+import com.affise.attribution.referrer.OnReferrerCallback
+import com.affise.attribution.referrer.ReferrerCallback
 
 
-class StatusModule : AffiseModule() {
+class StatusModule : AffiseModule(), ReferrerCallback {
 
     override val version: String = BuildConfig.AFFISE_VERSION
 
     private var checkStatusUseCase: CheckStatusUseCase? = null
+
+    private var referrerUseCase: ReferrerUseCase? = null
 
     override fun start() {
         val httpClient = get<HttpClient>()
@@ -45,9 +51,20 @@ class StatusModule : AffiseModule() {
             postBackModelFactory = postBackModelFactory,
             postBackConverter = postBackModelToJsonStringConverter,
         )
+
+        referrerUseCase = ReferrerUseCaseImpl(
+            checkStatusUseCase
+        )
     }
 
     override fun status(onComplete: OnKeyValueCallback) {
-        checkStatusUseCase?.send(onComplete) ?: onComplete.handle(emptyList())
+        checkStatusUseCase?.send { status ->
+            onComplete.handle(status)
+            referrerUseCase?.parseStatus(status)
+        } ?: onComplete.handle(emptyList())
+    }
+
+    override fun getReferrer(callback: OnReferrerCallback) {
+        referrerUseCase?.getReferrer(callback)
     }
 }