+++ module-status/src/main/java/com/affise/attribution/module/status/usecase/ReferrerUseCase.kt
@@ -1,0 +1,11 @@
+package com.affise.attribution.module.status.usecase
+
+import com.affise.attribution.modules.AffiseKeyValue
+import com.affise.attribution.referrer.OnReferrerCallback
+
+internal interface ReferrerUseCase {
+
+    fun getReferrer(onComplete: OnReferrerCallback)
+
+    fun parseStatus(status: List<AffiseKeyValue>): String?
+}